+++ deploy/Chart/charts/ucp/ucp-config.yaml
@@ -45,6 +45,12 @@ planes:
     properties:
       kind: "AWS"
 
+identity:
+  authentication: default
+
+ucp:
+  kind: kubernetes
+
 # UCP Metrics Provider
 metricsProvider:
   prometheus:
