+++ pkg/armrpc/hostoptions/providerconfig.go
@@ -6,6 +6,7 @@
 package hostoptions
 
 import (
+	"github.com/project-radius/radius/pkg/sdk"
 	"github.com/project-radius/radius/pkg/telemetry/metrics/provider"
 	"github.com/project-radius/radius/pkg/ucp/dataprovider"
 	qprovider "github.com/project-radius/radius/pkg/ucp/queue/provider"
@@ -23,7 +24,7 @@ type ProviderConfig struct {
 	Server          *ServerOptions                      `yaml:"server,omitempty"`
 	WorkerServer    *WorkerServerOptions                `yaml:"workerServer,omitempty"`
 	MetricsProvider provider.MetricsProviderOptions     `yaml:"metricsProvider"`
-	UCP             UCPConfig                           `yaml:"ucp"`
+	UCP             sdk.UCPOptions                      `yaml:"ucp"`
 	Logging         ucplog.LoggingOptions               `yaml:"logging"`
 
 	// FeatureFlags includes the list of feature flags.
